Moved searchView into SearchBarView
